Renamed AutocompleteValidator to AutocompleteBulkValidator.
